Add hardware smoke-test workflow for motors and ToF bring-up

+++ firmware/include/motor_driver.h
@@ -1,10 +1,7 @@
 #pragma once
 
-// Placeholder motor interface for two-wheel differential drive.
-// Positive values should command forward wheel motion.
-// Negative values should command reverse wheel motion.
-// Implementation details will depend on the chosen motor driver hardware.
+#include <stdint.h>
 
 void initMotors();
-void setMotorSpeeds(float left, float right);
+void setMotorSpeeds(int left, int right);
 void stopMotors();
